requirements.txt
django
django-ninja
django-cors-headers
djangorestframework-simplejwt